LayoutTests/fast/scrolling/mac/momentum-animator-expected.txt

PASS window.scrollY > 150 is true
PASS successfullyParsed is true

TEST COMPLETE
